Updated schema methods; fixes #751
Note that the product does not support composite uniqueness constraints.
The only way to do something similar would be to use a "node key", which
is only available in Enterprise Edition.

This change fixes the signatures of the *_uniqueness_constraint methods
to be consistent with the functionality offered by the product.